server/src/TestingApp/urls.py
from django.contrib import admin
from django.urls import path, include
import debug_toolbar

urlpatterns = [
    path('', include('app.urls')),
    path('ems/', include('app_ems.urls')),
    path('api/', include('api.urls')),
    path('admin/', admin.site.urls),
    path('api-auth/', include('rest_framework.urls')),
    path('__debug__/', include(debug_toolbar.urls)),
]
