server/requirements.txt
appnope==0.1.2
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
decorator==5.0.9
Django==3.2.5
django-admin==2.0.1
django-cors-headers==3.10.0
django-debug-toolbar==3.2.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-livereload-server==0.3.2
django-six==1.0.4
djangorestframework==3.12.4
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
screen==1.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
xlwt==1.3.0
